requirements.txt
Flask==2.2.3
Pillow==9.4.0
numpy==1.24.2
pandas==2.0.1
opencv-python==4.7.0.72
folium==0.14.0
geopy==2.3.0
Flask-Cors==3.0.10
